Migrate naive LLM factories
