+++ simulation/core/factories/action_generators/comment/__init__.py
@@ -1,0 +1,5 @@
+"""Comment-specific action-generator factories."""
+
+from .naive_llm import create_naive_llm_comment_generator
+
+__all__ = ["create_naive_llm_comment_generator"]
