+++ simulation/core/factories/action_generators/follow/__init__.py
@@ -1,0 +1,5 @@
+"""Follow-specific action-generator factories."""
+
+from .naive_llm import create_naive_llm_follow_generator
+
+__all__ = ["create_naive_llm_follow_generator"]
