+++ simulation/core/factories/action_generators/like/__init__.py
@@ -1,0 +1,5 @@
+"""Like-specific action-generator factories."""
+
+from .naive_llm import create_naive_llm_like_generator
+
+__all__ = ["create_naive_llm_like_generator"]
